began setting up plotting in mcmc class
