+++ autofit/non_linear/search/abstract_search.py
@@ -848,39 +848,6 @@ class NonLinearSearch(AbstractFactorOptimiser, ABC):
         """
         return self._class_config[section][attribute_name]
 
-    def plot_results(self, samples):
-
-        def should_plot(name):
-            return conf.instance["visualize"]["plots_search"][self.__class__.__name__][name]
-
-        from autofit.non_linear.plot.samples_plotters import SamplesPlotter
-        from autofit.non_linear.plot.output import Output
-
-        plotter = SamplesPlotter(
-            samples=samples,
-            output=Output(path=self.paths.image_path / "search", format="png"),
-        )
-
-        if should_plot("corner"):
-            plotter.corner()
-
-
-        # if should_plot("corner"):
-        #     plotter.corner(
-        #         panelsize=3.5,
-        #         yticksize=16,
-        #         xticksize=16,
-        #         bins=20,
-        #         plot_datapoints=False,
-        #         plot_density=False,
-        #         fill_contours=True,
-        #         levels=(0.68, 0.95),
-        #         labelpad=0.02,
-        #         range=np.ones(samples.model.total_free_parameters) * 0.999,
-        #         label_kwargs={"fontsize": 24},
-        #     )
-
-
     def perform_update(
         self,
         model: AbstractPriorModel,
@@ -1165,4 +1132,4 @@ class NonLinearSearch(AbstractFactorOptimiser, ABC):
         return isinstance(other, NonLinearSearch) and self.__dict__ == other.__dict__
 
     def plot_results(self, samples):
-        pass
+        raise NotImplementedError
