+++ autofit/non_linear/search/mcmc/emcee/search.py
@@ -4,7 +4,6 @@ from typing import Dict, Optional
 import emcee
 import numpy as np
 
-from autoconf import conf
 from autofit.database.sqlalchemy_ import sa
 from autofit.mapper.model_mapper import ModelMapper
 from autofit.mapper.prior_model.abstract import AbstractPriorModel
@@ -15,8 +14,6 @@ from autofit.non_linear.search.mcmc.auto_correlations import AutoCorrelationsSet
 from autofit.non_linear.search.mcmc.auto_correlations import AutoCorrelations
 from autofit.non_linear.samples.sample import Sample
 from autofit.non_linear.samples.mcmc import SamplesMCMC
-from autofit.plot import EmceePlotter
-from autofit.non_linear.plot.output import Output
 
 
 class Emcee(AbstractMCMC):
@@ -346,24 +343,3 @@ class Emcee(AbstractMCMC):
             raise FileNotFoundError(
                 f"The file search_internal.hdf does not exist at the path {self.paths.search_internal_path}"
             )
-
-    def plot_results(self, samples):
-        def should_plot(name):
-            return conf.instance["visualize"]["plots_search"]["emcee"][name]
-
-        plotter = EmceePlotter(
-            samples=samples,
-            output=Output(path=self.paths.image_path / "search", format="png"),
-        )
-
-        if should_plot("corner"):
-            plotter.corner()
-
-        if should_plot("trajectories"):
-            plotter.trajectories()
-
-        if should_plot("likelihood_series"):
-            plotter.likelihood_series()
-
-        if should_plot("time_series"):
-            plotter.time_series()
